Test TaskController DeleteTask

+++ src/Controller/TaskController.php
@@ -67,6 +67,10 @@ class TaskController extends AbstractController
     public function deleteTask(Task $task)
     {
         $user = $this->getUser();
+        if ($user === null){
+            $this->addFlash('error', 'Impossible de supprimer la tache veuillez vous connecter.');
+            return $this->redirectToRoute('app_login');                
+        }
 
         $result = $this->taskService->deleteTask($task, $user);
